// src/app/vampire-dark-age/model/dark-ages-sheet/vampire-dark-ages/disciplines/VampireDADisciplines.ts
import { VampireDADisciplineLevel } from "./VampireDADisciplineLevel";
import { VampireDAMagicDisciplineLevel } from "./VampireDAMagicDisciplineLevel";

export interface VampireDADisciplines {
  animalism: VampireDADisciplineLevel;
  auspex: VampireDADisciplineLevel;
  celerity: VampireDADisciplineLevel;
  dementation: VampireDADisciplineLevel;
  dominate: VampireDADisciplineLevel;
  quietus: VampireDADisciplineLevel

  fortitude: VampireDADisciplineLevel;
  mortis: VampireDAMagicDisciplineLevel;
  obtenebration: VampireDADisciplineLevel;
  obfuscate: VampireDADisciplineLevel;
  potence: VampireDADisciplineLevel;
  presence: VampireDADisciplineLevel;

  protean: VampireDADisciplineLevel;
  chimerstry: VampireDADisciplineLevel;
  serpentis: VampireDADisciplineLevel;
  thaumaturgy: VampireDAMagicDisciplineLevel;
  vicissitude: VampireDADisciplineLevel;

  temporis: VampireDADisciplineLevel;
  daimoinon: VampireDADisciplineLevel;
}